Move common generator processing functions outside the implementation directory.
