+++ src/lib/devices/Generator_common.h
@@ -12,6 +12,10 @@
  */
 
 
+#ifndef K_GENERATOR_COMMON_H
+#define K_GENERATOR_COMMON_H
+
+
 #include <stdlib.h>
 #include <stdbool.h>
 #include <stdio.h>
@@ -103,32 +107,6 @@ void Generator_common_handle_filter(
 
 
 /**
- * Process volume ramping at the start of a note.
- *
- * This function modifies WORK_BUFFER_AUDIO_* and should be called by generator
- * implementations (that need it) before returning from their process
- * function.
- *
- * \param gen          The Generator -- must not be \c NULL.
- * \param vstate       The Voice state -- must not be \c NULL.
- * \param wbs          The Work buffers -- must not be \c NULL.
- * \param ab_count     The number of audio buffers used -- must be \c 1 or
- *                     \c 2. If \c 1, WORK_BUFFER_AUDIO_L will be updated.
- * \param audio_rate   The audio rate -- must be positive.
- * \param buf_start    The start index of the buffer area to be processed.
- * \param buf_stop     The stop index of the buffer area to be processed.
- */
-void Generator_common_ramp_attack(
-        const Generator* gen,
-        Voice_state* vstate,
-        const Work_buffers* wbs,
-        int ab_count,
-        uint32_t audio_rate,
-        int32_t buf_start,
-        int32_t buf_stop);
-
-
-/**
  * Process volume ramping at the end of a note.
  *
  * This function modifies WORK_BUFFER_AUDIO_* and should be called after the
@@ -178,5 +156,8 @@ void Generator_common_handle_panning(
         const Work_buffers* wbs,
         int32_t buf_start,
         int32_t buf_stop);
+
+
+#endif // K_GENERATOR_COMMON_H
