+++ src/lib/devices/generators/Generator_debug.c
@@ -18,7 +18,7 @@
 
 #include <debug/assert.h>
 #include <devices/Device_params.h>
-#include <devices/generators/Generator_common.h>
+#include <devices/generators/Gen_utils.h>
 #include <devices/generators/Generator_debug.h>
 #include <memory.h>
 #include <player/Work_buffers.h>
