+++ src/lib/devices/param_types/Sample.c
@@ -20,7 +20,6 @@
 #include <math.h>
 
 #include <debug/assert.h>
-#include <devices/generators/Generator_common.h>
 #include <devices/param_types/Sample.h>
 #include <devices/param_types/Sample_mix.h>
 #include <devices/param_types/Sample_params.h>
